Fixing precedence for 'noop' notifies

+++ src/embr/observer.h
@@ -23,7 +23,7 @@ namespace internal {
 // Used so that on_notify calls are optional
 // fallback one for when we just can't match the on_notify
 template <class TObserver, class TEvent>
-static auto notify_helper(TObserver& observer, const TEvent& n, int) -> bool
+static auto notify_helper(TObserver& observer, const TEvent& n, long) -> bool
 {
     // TODO: Perhaps put our "didn't fire" warning here - compile time would be best
     return true;
@@ -32,7 +32,7 @@ static auto notify_helper(TObserver& observer, const TEvent& n, int) -> bool
 
 // fallback for invocation with context where no on_notify is present
 template <class TObserver, class TEvent, class TContext>
-static auto notify_helper(TObserver& observer, const TEvent& n, TContext&, int) -> bool
+static auto notify_helper(TObserver& observer, const TEvent& n, TContext&, long) -> bool
 {
     // TODO: Perhaps put our "didn't fire" warning here - compile time would be best
     return true;
@@ -53,7 +53,7 @@ static auto notify_helper(TObserver& observer, const TEvent& n, bool)
 // a little concerned, since this is preferred over the non-existing on_notify flavors, but the
 // decltype seems to step in and boost precedence
 template <class TObserver, class TEvent, class TContext>
-static auto notify_helper(TObserver& observer, const TEvent& n, TContext& context, long)
+static auto notify_helper(TObserver& observer, const TEvent& n, TContext& context, int)
     -> decltype(std::declval<TObserver>().on_notify(n), void(), bool{})
 {
     observer.on_notify(n);
